requirements.txt
pygame
numpy
matplotlib
pillow